requirements.txt
python-dotenv==1.1.0
openai==1.75.0
moviepy==2.1.2
requests==2.32.3
Flask==3.1.0